interface to libao2 changed ao_plugin added
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@3096 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libao2/audio_out_internal.h
@@ -7,7 +7,7 @@ static void uninit();
 static void reset();
 static int get_space();
 static int play(void* data,int len,int flags);
-static int get_delay();
+static float get_delay();
 static void audio_pause();
 static void audio_resume();
